Removed unnecessary files and Player Turn
Removed files that were made by URP. These files were just demo/example files and are not needed for the program.

The game now displays whose turn it is, added some code that changes text colour but does not work yet, but can be modified later when want to make everything nicer.